requirements.txt
Flask
numpy==1.26.4
scipy
pandas
gunicorn
# psycopg2-binary
Flask-SQLAlchemy
torch
joblib
scikit-learn
umap-learn
flask-socketio
